v0.4 "Chrono"
The roulette function is now regulated so as to make it's minimum spin still move a bit and it's highest spin to be done within a reasonable amount of time. This is done by limiting the velocity within a 10th it's normal range and boosting the minimum. That's the only big change I can remember, but still a good enough feature to warrant a new version.